Fixed create date in comment header.
This thing's been on GitHub for over three years, and I only just
realized that I let the auto-generated header use "today" as the created
date when I added it? No, this was made sometime around November
'08--February '09, and the header was added in '13. Geez, so much wrong
before I even hit the code. I was just skimming over, but now I really
want to go back and fix this. So much to fix. I might toss in some STB
stuff to make the code cleaner, or I might do everything in straight
standard-library C.