Moved desktop version into the repo and switched from Makefile to build.sh
I wrote the desktop version from the laptop version, but in 2012 or 13,
and it didn't need anywhere near as much improvement. I cleaned up the
comments and variable names before committing it. I also switched from
Makefile (which is just annoying to use) to a .sh build file (which is
easy when you already use bash scripts all the time). The default build
is laptop, but I haven't cleaned up the laptop version yet. I've begun to
think about making this cross-platform, since I do spend a fair amount of
time in the terminal when I'm on Windows as well. Obviously, that'll
require I think a bit more about the build chain and ACPI drivers.