A little cleanup, plus a little README editing.
README might still need a little editing; it's late, and I'm still not
too committed to fixing that up. Markdown is just close enough to HTML to
confuse me sometimes, so it's kind of annoying to work with. On the other
hand, at this point, I'm willing to call this version 2.0. 1.0 was the
crappy thing I just tossed up here without really editing; I'll cop to my
mistakes.