Quick backwards-compatibility hack.
Which I can't even test, because it's only for the battery info, and my
laptop uses the newer kernel version this works for. My desktop's kernel
is old enough, but it's, well, a desktop. I don't think I could make a
VirtualBox VM that could track battery life either. Well, _maybe_ I could
do it on the laptop, and use the laptop's battery as the VM's battery,
but 1) that old thing would probably choke on VirtualBox, and 2) that
sounds like the hacky sort of thing that won't actually work anyway.